Rewrote the Carpenter.cs to be CarpenterYoung.cs, so every age will have a different file. Changed the emotion state to not work with dispositions (not the prettiest, but it is easy to follow). Emotion states are broken up into low, med, high disposition, and then special cases. I added in an extra constructor that will take the current dialogue and keep it in the new state.
Gave NPC.cs some static variables for disposition use. Disposition is updated with a delta instead of changed to a specific number.

Set up a default dialogue for items that the NPC does not want. This is a dictionary that is called and returns a string for the npc to then say. Added in some non item specific keys that are in Strings.cs.